Anon no longer defaults to None
Must now be explicitly specified True or False - no attempt to use
credentials with fallback. No check of successful credentials is made,
error will only happen when an operation is attempted.

If not explicitly provided, but not anonymous, delegation token will
not call STS, but assume that whatever mechanism allowed

Changed docs to reflect these changes, with a short section on
credential usage