ensure multi-parts written by ``flush()`` respect ``block_size`` and S3 max part size
1) ``flush()`` was properly ensuring that the buffer had reached at
least ``block_size`` in length before writing a part to S3; however, it
was not ensuring that resulting part(s) were themselves at all on the
order of ``block_size``.

2) Nor did the method ensure every part is less than S3's maximum part
size, (regardless of desired ``block_size``).

---

(1) was relatively "easy" to cross -- if a single ``write()`` payload
was of *any* size equal to or greater than ``buffer_size`` -- even many
times larger than ``buffer_size`` -- it was uploaded as is, as a single
part. This itself caused ``flush()`` to violate the expectation of
``buffer_size``, and to operate differently depending on whether the
user wrote their payload in small chunks or big chunks.

Moreover, (1) caused a breaking issue, when a single write payload was
greater than S3's maximum upload (part) limit of 5 GiB. This is a large
string to write, certainly; however, not an unreasonable one, when your
objects are that large -- (and, S3 objects can be much larger than
that).

(1) is fixed by making ``flush()`` read and upload ``block_size`` chunks
from the buffer, (only writing a larger part than that at the end of the
buffer, in order to avoid a smaller-than-``buffer_size`` part).

---

(2) is even more of an edge case, but likely worth handling, at the same
time -- while ``buffer_size`` is (hypothetically) configurable, and up
to the configurer to ensure that it is neither too small nor too large,
so long as ``flush()`` is ever writing parts *larger* than
``buffer_size``, then it is still conceivable for it to request to
upload a part that is larger than S3's maximum. However, this
implies that ``buffer_size`` has been configured to be far larger than
S3's *minimum* -- really, it must be more than half as large as the
maximum -- and therefore ``buffer_size`` may be safely, and trivially,
violated, by uploading the remainder of the buffer in two equal-size
parts, (and satisfying the part size minimum for both).